updating encodings
